+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # CHANGELOG
 
 ## unreleased -- v0.9.2
+<<<<<<< HEAD
 ### Added
 - `readerVersion` parameter for `/search`, `/work` and `/edition` endpoints to control media types returned
+=======
+### Fixed
+- Improve clustering stability by improving individual error handling
+>>>>>>> main
 
 ## 2021-09-09 -- v0.9.1
 ### Fixed